Allow internal persistent auxiliary field output
This commit adds the capability for WarpX to output fields (multifabs)
that are registered in the multifab-register and are persistent at the
time of output.
Users will also be able to create fields via Python callback functions
and add them to the normal output.

The check in picmi.py for matching strings now looks outdated and should
perhaps be removed at a later point in time.