Add callback beforediagnostics and documentation
This commit adds another callback that allows for code to be executed
directly before full diagnostics are written.
It also adds extensive documentation on how to use the new features to
expose multifabs in WarpX to diagnostics, or how to create new multifabs
in a kernel that can be added to regular output.