Add Hall and Pressure Terms to Diags
Hard-coded Hall- and pressure terms in hybrid solver to diagnostics.
These are the terms that, when divided by the density (in the algorithm
limited from below by n_floor) will result in the electric field in the
Ohm's law.

E.g. Ex(i, j, k) = (enE_x - grad_Pe) / rho_val_limited

But to get these, a user needs to allocated the fields via the Python
interface, and the 3D Ohm Solver cylinder compression test shows how
that is done.